# ANSI parser speedup