--- a/process/ansi.go
+++ b/process/ansi.go
@@ -8,7 +8,7 @@ package process
 // including mention of some of the necessary deviations from the standards
 // (such as allowing some sequences to terminate with BEL instead of ESC '\').
 type ansiParser struct {
-	state ansiParserState
+	state *ansiParserState
 }
 
 // Write passes more bytes through the parser.
@@ -31,11 +31,11 @@ func (m *ansiParser) insideCode() bool { return m.state != nil }
 
 // ansiParserState is a possible state of the parser. It's a map of incoming-
 // byte to next-state. Most next-states are nil (they exit the escape code).
-type ansiParserState map[byte]ansiParserState
+type ansiParserState [256]*ansiParserState
 
 var (
 	// initialANSIState is the state the parser enters once it reads ESC.
-	initialANSIState = ansiParserState{
+	initialANSIState = &ansiParserState{
 		// Note that most bytes immediately following ESC terminate the sequence.
 		// The following require more processing:
 		'[': csiParameterState, // CSI
@@ -46,10 +46,10 @@ var (
 		'_': stTextState,       // APC
 	}
 	// csiParameter state is the state the parser is in after ESC '['
-	csiParameterState = ansiParserState{}
+	csiParameterState = &ansiParserState{}
 
 	// stTextState is one of the ST-terminated text states (OSC, DCS, APC, etc)
-	stTextState = ansiParserState{}
+	stTextState = &ansiParserState{}
 )
 
 // The "looping states" can't be built as struct literals since they refer to
@@ -72,7 +72,7 @@ func init() {
 	//             |
 	//             +--anything else--> (nil)
 	//
-	csiIntermediate := ansiParserState{}
+	csiIntermediate := &ansiParserState{}
 	for b := byte(0x30); b <= 0x3F; b++ {
 		csiParameterState[b] = csiParameterState
 	}
@@ -95,7 +95,7 @@ func init() {
 	//                  |
 	//                  +--BEL--> (nil)
 	//
-	stEscapeState := ansiParserState{}
+	stEscapeState := &ansiParserState{}
 	for b := range 256 {
 		stTextState[byte(b)] = stTextState
 		stEscapeState[byte(b)] = stTextState