… process/fixtures/npm.sh.raw