### Description

Significant speed savings:

```
❯ benchstat ansi-parser-{maps,arrays}.txt
goos: darwin
goarch: arm64
pkg: github.com/buildkite/agent/v3/process
cpu: Apple M1 Max
              │ ansi-parser-maps.txt │       ansi-parser-arrays.txt        │
              │        sec/op        │   sec/op     vs base                │
ANSIParser-10           2322.3µ ± 0%   565.3µ ± 1%  -75.66% (p=0.000 n=30)
```

### Context

The secret redactor was sped up a bit by switching away from a map, and I remembered this used maps too.

### Changes

* Add a benchmark
* Rename `feed` to `Write` in case we want to e.g. `io.Copy` or `MultiWriter` into it
* Use `[256]*ansiParserState` instead of `map[byte]ansiParserState`

### Testing
- [x] Tests have run locally (with `go test ./...`). Buildkite employees may check this if the pipeline has run automatically.
- [x] Code is formatted (with `go fmt ./...`)